Split and rename SkipIfMissingRequiredImage
This function skips if the PV is not exists AND modifies the PV to
remove its claimRef field. This is confusing.

This commit splits the function so now it only skip if the pv does not
exist, and returns the PV. It also rename the function to reflect that.
The new name is `GetRequiredImageAndSkipIfMissing`

The commit also adds a new `MakePVAvailable` function to libstorage and
moves all pv function to a new pv.go file.

The function uses the patch instead of update, to modify the PV.
